--- src/view/screens/Search/index.tsx
@@ -1,1 +1,0 @@
-export {SearchScreen} from '#/view/screens/Search/Search'
